+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added `SyncFormRequest`, `SyncFormEntryRequest` and `SyncAccount`.
 
 ## [0.6.0] - 2019-01-03
 ### Added